Generalize ajax form stimulus controller and implement new course modal; fix rounding for course distances; move staging breadcrumbs to partial.

+++ app/javascript/controllers/event_form_controller.js
@@ -23,7 +23,7 @@ export default class extends Controller {
 
     fillDistance() {
         const select = this.courseDropdownTarget;
-        const courseDistance = parseInt(select.options[select.selectedIndex].dataset.distance);
+        const courseDistance = parseFloat(select.options[select.selectedIndex].dataset.distance);
         const distanceUnit = this.formDataTarget.dataset.prefDistanceUnit;
         const laps = parseInt(this.lapDropdownTarget.value);
